chore: moved expert types to shared module and separated document status into its own enum
