+++ src/core/experts/ExpertManager.ts
@@ -2,7 +2,7 @@ import fs from "fs/promises"
 import * as path from "path"
 import * as vscode from "vscode"
 import { v4 as uuidv4 } from "uuid"
-import { DocumentLink, ExpertData, ExpertDataSchema } from "../../../webview-ui/src/types/experts"
+import { DocumentLink, ExpertData, ExpertDataSchema } from "../../shared/experts"
 import { fileExistsAtPath, createDirectoriesForFile } from "../../utils/fs"
 import { GlobalFileNames } from "../../global-constants"
 import { UrlContentFetcher } from "../../services/browser/UrlContentFetcher"
